+++ packages/core-components/package.json
@@ -13,11 +13,9 @@
     "clsx": "^1.2.1",
     "react-icons": "^4.9.0",
     "styled-components": "^6.0.0-rc.3",
-    "tailwind-merge": "^1.13.2"
-  },
-  "peerDependencies": {
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0"
+    "tailwind-merge": "^1.13.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@rollup/plugin-babel": "^6.0.3",
