+++ playground/src/components/buttons/Select.stories.tsx
@@ -7,8 +7,8 @@ import { GiUsaFlag } from "react-icons/gi";
 
 import "regenerator-runtime/runtime";
 import "@namviet-fe/core-ui/dist/style.css";
-import EnIcon from "../assets/buttons/en";
-import ViIcon from "../assets/buttons/vi";
+import EnIcon from "../../assets/buttons/en";
+import ViIcon from "../../assets/buttons/vi";
 export default {
   title: "@nv-fe/core-ui/buttons/Select",
 } satisfies StoryDefault<DropdownButtonProps>;
